Move elaboration order code into sub-module of elaboration
