+++ fathom/src/surface.rs
@@ -3,7 +3,6 @@
 use std::cell::RefCell;
 
 use codespan_reporting::diagnostic::{Diagnostic, Label};
-use fxhash::{FxHashMap, FxHashSet};
 use lalrpop_util::lalrpop_mod;
 use scoped_arena::Scope;
 
@@ -43,301 +42,6 @@ impl<'arena> Module<'arena, ByteRange> {
             });
 
         (term, messages)
-    }
-
-    pub fn elaboration_order(&self) -> Result<Vec<usize>, Vec<ModuleOrderMessage>> {
-        let item_names = self.item_names();
-        let item_deps = self.item_dependencies(&item_names);
-
-        let mut context = ModuleOrderContext::new();
-        context.determine_order(&self.items, &item_names, &item_deps);
-
-        if context.messages.is_empty() {
-            Ok(context.output)
-        } else {
-            Err(context.messages)
-        }
-    }
-
-    fn item_names(&self) -> FxHashMap<StringId, usize> {
-        self.items
-            .iter()
-            .enumerate()
-            .filter_map(|(i, item)| match item {
-                Item::Definition {
-                    label: (_, label), ..
-                } => Some((*label, i)),
-                Item::ReportedError(_) => None,
-            })
-            .collect()
-    }
-
-    fn item_dependencies(&self, item_names: &FxHashMap<StringId, usize>) -> Vec<Vec<StringId>> {
-        let mut local_names = Vec::new();
-        self.items
-            .iter()
-            .map(|item| item_dependencies(item, &item_names, &mut local_names))
-            .collect()
-    }
-}
-
-struct ModuleOrderContext {
-    output: Vec<usize>,
-    visited: FxHashSet<StringId>,
-    stack: Vec<StringId>,
-    messages: Vec<ModuleOrderMessage>,
-}
-
-impl ModuleOrderContext {
-    fn new() -> Self {
-        ModuleOrderContext {
-            output: Vec::new(),
-            visited: FxHashSet::default(),
-            stack: Vec::new(),
-            messages: Vec::new(),
-        }
-    }
-
-    fn determine_order(
-        &mut self,
-        items: &[Item<'_, ByteRange>],
-        item_names: &FxHashMap<StringId, usize>,
-        dependencies: &[Vec<StringId>],
-    ) {
-        for item in items {
-            match item {
-                Item::Definition {
-                    label: (_, label), ..
-                } => {
-                    self.visit_item(*label, item_names, dependencies);
-                    self.stack.clear();
-                }
-                Item::ReportedError(_) => {}
-            }
-        }
-    }
-
-    fn visit_item(
-        &mut self,
-        name: StringId,
-        item_names: &FxHashMap<StringId, usize>,
-        dependencies: &[Vec<StringId>],
-    ) {
-        if self.visited.contains(&name) {
-            return;
-        }
-
-        if self.stack.contains(&name) {
-            self.stack.push(name);
-            self.messages.push(ModuleOrderMessage::CycleDetected {
-                names: self.stack.clone(),
-            });
-            return;
-        }
-
-        match item_names.get(&name) {
-            Some(index) => {
-                let index = *index;
-                self.stack.push(name);
-                let deps = &dependencies[index];
-                for dep in deps {
-                    self.visit_item(*dep, item_names, dependencies);
-                }
-                self.stack.pop();
-                self.visited.insert(name);
-                self.output.push(index);
-            }
-            // not a module item
-            None => {}
-        }
-    }
-}
-
-fn item_dependencies(
-    item: &Item<'_, ByteRange>,
-    item_names: &FxHashMap<StringId, usize>,
-    local_names: &mut Vec<StringId>,
-) -> Vec<StringId> {
-    let mut deps = Vec::new();
-    match item {
-        Item::Definition {
-            label: _,
-            type_,
-            expr,
-        } => {
-            if let Some(r#type) = type_ {
-                term_deps(r#type, item_names, local_names, &mut deps);
-            }
-            term_deps(expr, item_names, local_names, &mut deps);
-        }
-        Item::ReportedError(_) => {}
-    }
-    deps
-}
-
-fn term_deps(
-    term: &Term<ByteRange>,
-    item_names: &FxHashMap<StringId, usize>,
-    local_names: &mut Vec<StringId>,
-    deps: &mut Vec<StringId>,
-) {
-    match term {
-        Term::Name(_, name) => {
-            if local_names.iter().rev().any(|local| name == local) {
-                // local binding, do nothing
-            } else if item_names.contains_key(name) {
-                if deps.last() != Some(name) {
-                    // Only push if it's not a duplicate of the last item. This is a basic way
-                    // to reduce the number of duplicate dependencies that are pushed.
-                    deps.push(*name);
-                }
-            }
-        }
-        Term::Ann(_, expr, r#type) => {
-            term_deps(expr, item_names, local_names, deps);
-            term_deps(r#type, item_names, local_names, deps);
-        }
-        Term::Let(_, pattern, r#type, def_expr, body_expr) => {
-            push_pattern(pattern, local_names);
-            if let Some(r#type) = r#type {
-                term_deps(r#type, item_names, local_names, deps);
-            }
-            term_deps(def_expr, item_names, local_names, deps);
-            term_deps(body_expr, item_names, local_names, deps);
-            pop_pattern(pattern, local_names);
-        }
-        Term::Match(_, scrutinee, equations) => {
-            let initial_locals_names_len = local_names.len();
-            term_deps(scrutinee, item_names, local_names, deps);
-            for (pattern, body) in *equations {
-                push_pattern(pattern, local_names);
-                term_deps(body, item_names, local_names, deps);
-            }
-            local_names.truncate(initial_locals_names_len);
-        }
-        Term::Arrow(_, input_type, output_type) => {
-            term_deps(input_type, item_names, local_names, deps);
-            term_deps(output_type, item_names, local_names, deps);
-        }
-        Term::FunType(_, input_pattern, input_type, output_type) => {
-            push_pattern(input_pattern, local_names);
-            if let Some(input_type) = input_type {
-                term_deps(input_type, item_names, local_names, deps);
-            }
-            term_deps(output_type, item_names, local_names, deps);
-            pop_pattern(input_pattern, local_names);
-        }
-        Term::FunLiteral(_, input_param, input_type, output_type) => {
-            push_pattern(input_param, local_names);
-            if let Some(input_type) = input_type {
-                term_deps(input_type, item_names, local_names, deps);
-            }
-            term_deps(output_type, item_names, local_names, deps);
-            pop_pattern(input_param, local_names);
-        }
-        Term::App(_, head_expr, input_expr) => {
-            term_deps(head_expr, item_names, local_names, deps);
-            term_deps(input_expr, item_names, local_names, deps);
-        }
-        Term::RecordType(_, type_fields) => {
-            let initial_locals_names_len = local_names.len();
-            for type_field in *type_fields {
-                term_deps(&type_field.type_, item_names, local_names, deps);
-                local_names.push(type_field.label.1);
-            }
-            local_names.truncate(initial_locals_names_len);
-        }
-        Term::RecordLiteral(_, expr_fields) => {
-            let initial_locals_names_len = local_names.len();
-            for expr_field in *expr_fields {
-                term_deps(&expr_field.expr, item_names, local_names, deps);
-                local_names.push(expr_field.label.1);
-            }
-            local_names.truncate(initial_locals_names_len);
-        }
-        Term::Proj(_, head_expr, _) => {
-            term_deps(head_expr, item_names, local_names, deps);
-        }
-        Term::ArrayLiteral(_, terms) => {
-            for term in *terms {
-                term_deps(term, item_names, local_names, deps);
-            }
-        }
-        Term::FormatRecord(_, fields) => {
-            field_deps(fields, item_names, local_names, deps);
-        }
-        Term::FormatOverlap(_, format_fields) => {
-            field_deps(format_fields, item_names, local_names, deps);
-        }
-        Term::FormatCond(_, (_, name), format, cond) => {
-            local_names.push(*name);
-            term_deps(format, item_names, local_names, deps);
-            term_deps(cond, item_names, local_names, deps);
-            local_names.pop();
-        }
-        Term::Hole(_, _)
-        | Term::Placeholder(_)
-        | Term::Universe(_)
-        | Term::UnitLiteral(_)
-        | Term::StringLiteral(_, _)
-        | Term::NumberLiteral(_, _)
-        | Term::BooleanLiteral(_, _)
-        | Term::ReportedError(_) => {}
-    }
-}
-
-fn field_deps(
-    fields: &[FormatField<ByteRange>],
-    item_names: &FxHashMap<StringId, usize>,
-    local_names: &mut Vec<StringId>,
-    deps: &mut Vec<StringId>,
-) {
-    let initial_locals_names_len = local_names.len();
-    for field in fields {
-        match field {
-            FormatField::Format {
-                label: (_, label),
-                format,
-                ..
-            } => {
-                term_deps(format, item_names, local_names, deps);
-                local_names.push(*label)
-            }
-            FormatField::Computed {
-                label: (_, label),
-                type_,
-                expr,
-            } => {
-                if let Some(r#type) = type_ {
-                    term_deps(r#type, item_names, local_names, deps);
-                }
-                term_deps(expr, item_names, local_names, deps);
-                local_names.push(*label)
-            }
-        }
-    }
-    local_names.truncate(initial_locals_names_len);
-}
-
-fn push_pattern(pattern: &Pattern<ByteRange>, local_names: &mut Vec<StringId>) {
-    match pattern {
-        Pattern::Name(_, name) => local_names.push(*name),
-        Pattern::Placeholder(_) => {}
-        Pattern::StringLiteral(_, _) => {}
-        Pattern::NumberLiteral(_, _) => {}
-        Pattern::BooleanLiteral(_, _) => {}
-    }
-}
-
-fn pop_pattern(pattern: &Pattern<ByteRange>, local_names: &mut Vec<StringId>) {
-    match pattern {
-        Pattern::Name(_, _) => {
-            local_names.pop();
-        }
-        Pattern::Placeholder(_) => {}
-        Pattern::StringLiteral(_, _) => {}
-        Pattern::NumberLiteral(_, _) => {}
-        Pattern::BooleanLiteral(_, _) => {}
     }
 }
 
@@ -695,29 +399,6 @@ fn format_expected(expected: &[impl std::fmt::Display]) -> Option<String> {
         (last, []) => format!("expected {}", last),
         (last, expected) => format!("expected {} or {}", expected.iter().format(", "), last),
     })
-}
-
-#[derive(Clone, Debug)]
-pub enum ModuleOrderMessage {
-    CycleDetected { names: Vec<StringId> },
-}
-
-impl ModuleOrderMessage {
-    pub fn to_diagnostic(&self, interner: &RefCell<StringInterner>) -> Diagnostic<FileId> {
-        match self {
-            ModuleOrderMessage::CycleDetected { names } => {
-                let interner = interner.borrow();
-                let names: Vec<_> = names
-                    .iter()
-                    .map(|id| interner.resolve(*id).unwrap())
-                    .collect();
-                let cycle = names.join(" → ");
-                Diagnostic::error()
-                    .with_message("cycle detected")
-                    .with_notes(vec![cycle])
-            }
-        }
-    }
 }
 
 #[cfg(test)]
